requirements.txt
inquirer==3.1.3
tabulate==0.9.0
requests==2.31.0
networkx==3.2.1
matplotlib==3.8.1
python-dotenv==1.0.0
paramiko==3.3.1
prettytable==3.9.0